more corrections
